refactor(ui): 修正 login 與 register 頁面程式碼
修正 login 與 register 頁面程式碼
