+++ restaurant_recommendation_system/mobile_app/app/register.tsx
@@ -1,10 +1,10 @@
 import React, { useState } from 'react';
 import { View, Text, TouchableOpacity, SafeAreaView, Alert } from 'react-native';
 import { useRouter, Link } from 'expo-router';
-import CustomInput from '../src/components/CustomInput';
-import CustomButton from '../src/components/CustomButton';
-import { register } from '../src/api/auth';
-import { sharedStyles } from '../src/styles/sharedStyles';
+import CustomInput from '../components/CustomInput';
+import CustomButton from '../components/CustomButton';
+import { register } from '../api/auth';
+import { globalStyles } from '../constants/styles';
 
 const RegisterScreen = () => {
   const [username, setUsername] = useState('');
@@ -33,31 +33,34 @@ const RegisterScreen = () => {
   };
 
   return (
-    <SafeAreaView style={sharedStyles.container}>
+    <SafeAreaView style={globalStyles.container}>
       <View>
-        <Text style={sharedStyles.title}>建立新帳號</Text>
-        {error ? <Text style={sharedStyles.errorText}>{error}</Text> : null}
+        <Text style={globalStyles.title}>建立新帳號</Text>
+        {error ? <Text style={globalStyles.errorText}>{error}</Text> : null}
         <CustomInput
           value={username}
           onChangeText={setUsername}
           placeholder="使用者名稱"
+          style={globalStyles.input}
         />
         <CustomInput
           value={email}
           onChangeText={setEmail}
           placeholder="電子郵件"
           keyboardType="email-address"
+          style={globalStyles.input}
         />
         <CustomInput
           value={password}
           onChangeText={setPassword}
           placeholder="密碼"
           secureTextEntry
+          style={globalStyles.input}
         />
-        <CustomButton title="註冊" onPress={handleRegister} />
+        <CustomButton title="註冊" onPress={handleRegister}style={globalStyles.button} textStyle={globalStyles.buttonText} />
         <Link href="/login" asChild>
             <TouchableOpacity>
-                <Text style={sharedStyles.linkText}>已經有帳號了？點此登入</Text>
+                <Text style={globalStyles.linkText}>已經有帳號了？點此登入</Text>
             </TouchableOpacity>
         </Link>
       </View>
